已完成： 已测试完TeachInfoController,TeacherController内容
代办：
测试CommentTeacherController内容
service层。（处理视频，音频）
管理模块

+++ src/main/java/Evaluation/dao/TeachInfoDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
  * Created by Lee on 2017/11/7 0007.
  */
 @Repository
-public interface TeachInfoDao extends JpaRepository<TeachInfo,Long> {
+public interface TeachInfoDao extends JpaRepository<TeachInfo,Integer> {
 
 
     TeachInfo findById(Integer integer);
